requirements.txt
altgraph==0.17.4
blinker==1.9.0
bottle==0.13.3
cffi==1.17.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
Counter==1.0.0
et_xmlfile==2.0.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pefile==2023.2.7
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pytz==2025.2
pywebview==5.4
pywin32-ctypes==0.2.3
setuptools==80.2.0
six==1.17.0
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
